Title: 시력 검사 2, Time: 165ms, Memory: 11MB, Status: Passed - CodeTree

+++ 231127/시력 검사 2/eye-test-2.java
@@ -3,7 +3,7 @@ import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
         Scanner input = new Scanner(System.in);
-        double vision = input.nextInt();
+        double vision = input.nextDouble();
         if (vision >= 1.0) {
             System.out.println("High");
         }
